+++ src/index.js
@@ -1,4 +1,4 @@
-export { stringifyInfo } from './stringify-info.js';
-export { stringifyStream } from './stringify-stream.js';
-export { stringifyChunked } from './stringify-chunked.js';
 export { parseChunked } from './parse-chunked.js';
+export { stringifyChunked } from './stringify-chunked.js';
+export { stringifyInfo } from './stringify-info.js';
+export { createStringifyWebStream, parseFromWebStream } from './web-streams.js';