+++ src/stringify-stream-browser.js
@@ -1,3 +1,0 @@
-export function stringifyStream() {
-    throw new Error('Method is not supported');
-};