+++ test/stringify-chunked.js
@@ -1,6 +1,6 @@
 import assert from 'node:assert';
 import { inspect } from 'node:util';
-import { stringifyChunked } from '../src/stringify-chunked.js';
+import { stringifyChunked } from '@discoveryjs/json-ext';
 import { wellformedStringify } from './helpers/well-formed-stringify.js';
 import { date, allUtf8LengthDiffChars, tests, spaceTests, spaces, replacerTests } from './fixture/stringify-cases.js';